Refactor import statements for visualization module
